# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: conpass API
  version: 0.1.0
  description: >
    Loyalty & membership platform API. Single source of truth (contract-first).
    Backend is the authority for balances; wallet passes are display-only.
    Each tag maps to one single-responsibility Lambda service under backend/services.
  contact:
    name: conpass engineering
servers:
  - url: https://api.conpass.cards
    description: Production (frontend is served from https://conpass.cards)
  - url: http://localhost:3000
    description: Local (serverless offline)
...
security:
  - supabaseJwt: []
paths:
...
  /programs:
    get:
      tags: [programs]
      operationId: listPrograms
      summary: List programs for the caller's merchant
      parameters:
        - $ref: "#/components/parameters/MerchantIdQuery"
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Program" }
    post:
      tags: [programs]
      operationId: createProgram
      summary: Create & publish a program (enforces tier program limit)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProgramCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "402": { $ref: "#/components/responses/TierLimit" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /programs/{programId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
    get:
      tags: [programs]
      operationId: getProgram
      summary: Get a program
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [programs]
      operationId: updateProgram
      summary: Update program parameters / appearance / active state
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProgramUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
...
components:
  securitySchemes:
    supabaseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth (GoTrue) access token, validated via JWKS.
  parameters:
...
    MerchantIdQuery:
      name: merchantId
      in: query
      required: false
      schema: { type: string, format: uuid }
    ProgramId:
      name: programId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TierLimit:
      description: Subscription tier limit reached
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine-readable error code }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    ProgramType:
      type: string
      enum: [loyalty_stamps, loyalty_points, membership_pass]
    RewardMechanic:
      type: string
      enum: [stamps, points]
    MembershipValidity:
      type: string
      enum: [monthly, quarterly, annual, per_event]
...
    ProgramAppearance:
      type: object
      properties:
        color: { type: string, description: Hex accent color. }
        iconStorageKey: { type: string, description: 512x512 PNG. }
        backgroundStorageKey: { type: string, description: 1032x336 JPG. }
      additionalProperties: false
    ProgramCreate:
      type: object
      required: [type, name, appearance, wallets]
      properties:
        type: { $ref: "#/components/schemas/ProgramType" }
        name: { type: string }
        mechanic: { $ref: "#/components/schemas/RewardMechanic" }
        stampsForReward: { type: integer, minimum: 1 }
        pointsForReward: { type: integer, minimum: 1 }
        pointsPerDollar: { type: number, minimum: 0 }
        reward: { type: string, description: Reward description shown on the card. }
        membershipValidity: { $ref: "#/components/schemas/MembershipValidity" }
        membershipIncludes: { type: string }
        welcomeBonus: { type: integer, default: 0 }
        expiryDays: { type: integer, description: Balance expiry window; null = no expiry. }
        appearance: { $ref: "#/components/schemas/ProgramAppearance" }
        wallets:
          type: array
          items: { type: string, enum: [apple, google] }
          minItems: 1
      additionalProperties: false
    ProgramUpdate:
      type: object
      properties:
        name: { type: string }
        reward: { type: string }
        stampsForReward: { type: integer, minimum: 1 }
        pointsForReward: { type: integer, minimum: 1 }
        pointsPerDollar: { type: number, minimum: 0 }
        appearance: { $ref: "#/components/schemas/ProgramAppearance" }
        active: { type: boolean }
      additionalProperties: false
    Program:
      allOf:
        - $ref: "#/components/schemas/ProgramCreate"
        - type: object
          required: [id, merchantId, active, createdAt]
          properties:
            id: { type: string, format: uuid }
            merchantId: { type: string, format: uuid }
            active: { type: boolean }
            enrollmentUrl: { type: string, format: uri }
            createdAt: { type: string, format: date-time }
...